cosiñas

+++ Dtos/OrdenCompra/DetalleOrdenCompraDto.cs
@@ -4,7 +4,7 @@
     {
         public long nDetalleOrdenCompra { get; set; }
         public long cantidadArticulos { get; set; }
-        public double precioSubTotal { get; set; }
+        public decimal precioSubTotal { get; set; }
         public long? idArticulo { get; set; }
         public string? nombreArticulo { get; set; } = "";
     }
